refactor(console): hands-free controller; dictation's exception retired (wave 2)
Extracts the V3 pipeline hands-free conversation loop into
ConsoleHandsFreeController (UI/Console_Modules/hands_free.py), the third
wave-2 controller. 28 methods (~730 lines) moved verbatim, plus the
engine fork (_enter_console_hands_free_loop) and both cross-engine action
entry points (action_toggle/action_exit_console_hands_free, check_action's
availability branch) it shares with PR #1350's V4 realtime loop.

Two-engine boundary: the controller owns the fork and both action entry
points (they read both engines = coordination logic); the realtime
engine's own ~60-method implementation stays on ChatScreen exactly as
#1350 wrote it, reached through two named callables
(realtime_session_accessor, enter_realtime_loop). The reverse direction
(realtime code calling the moved pipeline engine, e.g. the loud
fallback-to-pipeline path) is ordinary screen-calls-its-own-controller
traffic via self._hands_free, no injection needed.

Closes out wave 1's disclosed temporary exception in
ConsoleDictationController: its four hands-free reach-backs (plus
_console_realtime_adopt_transcript and _run_pending_console_voice_action,
same shape/reason) are now named keyword-only constructor callables
instead of live properties reaching through screen, matching the binding
rule uniformly. Both module docstrings drop the "temporary exception"
language.

ChatScreen keeps thin delegations/proxies for the handful of entry points
and state reached from outside the cluster (on_key, on_button_pressed,
on_unmount, tests): action_toggle/exit_console_hands_free,
_enter_console_hands_free_loop, and read/write properties for
_console_hands_free, _console_hands_free_vad_degraded,
_console_hands_free_store_tap_installed.

Verification: mechanical AST diff confirms all 28 moved methods are gone
from ChatScreen (zero dead bodies); pyflakes on chat_screen.py is
byte-identical to the pre-move baseline (module-level rename of
acoustic_barge_in_enabled/ExitLoop/ModeChanged/SilenceSpeech to dual
imports, handsfree_send_delay_seconds/resolve_handsfree_engine/
realtime_enabled/console_voice_input removed as newly-dead); pyflakes on
the new module is clean. Two test-helper traps fixed (config readers and
console_voice_input pinned on chat_screen_module now also/instead pinned
on the new hands_free module, since the pipeline engine reads its own
copy) plus 39 direct method-poke call sites retargeted from console.X to
console._hands_free.X.

Tests: 156 passed (hands-free + dictation-streaming + shell-regions,
byte-identical to a pre-move baseline run in a disposable worktree at
HEAD), 110 passed (realtime wiring + realtime loop FSM, proving #1350's
feature wasn't silently broken), 16 passed (dictation's own suite).
Zero regressions.
